test(steps): recategorize and clean up browser tests
